[scanner] fix: split CardWrapper and ConsoleOfflineDetectionCard components
- Extract CardWrapper.types.ts: CardContainerSize, CardExpandedContextType,
  CardFlashType, PendingSwap, CardWrapperProps interfaces
- Extract CardWrapper.constants.ts: LARGE_EXPANDED_CARDS, FULLSCREEN_EXPANDED_CARDS,
  timing constants, and CSS container query style
- CardWrapper.tsx reduced from 875 to 757 lines; re-exports types for compat
- Extract useOfflineDetection.ts hook from ConsoleOfflineDetectionCard.tsx:
  separates all data-fetching, state management, and computed values
- ConsoleOfflineDetectionCard.tsx reduced from 894 to 256 lines (renderer only)
- Remove eslint-disable max-lines comments from both files


# Conflicts:
#	web/src/components/cards/CardWrapper.tsx
